seperated the Horoscope Provider and the Card and updated the useState from setState
